refactor(renderer): 补全模块提取并修复 lazy 失效与 TS 类型问题
