+++ app/renderer/src/main/src/components/MilkdownEditor/CustomFile/CustomFile.tsx
@@ -29,11 +29,14 @@ import { YakitHint } from '@/components/yakitUI/YakitHint/YakitHint'
 import React from 'react'
 import { SolidCloudDownloadIcon } from '@/assets/newIcon'
 import useDownloadUrlToLocalHooks, { type DownloadUrlToLocal } from '@/hook/useDownloadUrlToLocal/useDownloadUrlToLocal'
+import { apiDownloadStorageType } from '@/pages/notepadManage/notepadStorageUtils'
 import {
-  apiDownloadStorageType,
   onOpenLocalFileByPath,
   saveDialogAndGetLocalFileInfo,
 } from '@/pages/notepadManage/notepadManage/utils'
+import { getTypeAndNameByPath } from './customFilePathUtils'
+
+export { getTypeAndNameByPath } from './customFilePathUtils'
 import type { YakitHintProps } from '@/components/yakitUI/YakitHint/YakitHintType'
 import useUploadOSSHooks, { type UploadFileTypeProps, type UploadOSSStartProps } from '@/hook/useUploadOSS/useUploadOSS'
 import { getHttpFileLinkInfo, getLocalFileLinkInfo } from './utils'
@@ -61,17 +64,6 @@ interface DownFileInfoProps {
   fileName: string
 }
 /**上传文件后,后端拼接hash和文件名的字符:&*&;方便截取文件名 */
-export const getTypeAndNameByPath = (path) => {
-  let newPath = decodeURIComponent(path.split('/').pop())
-  const firstIndex = newPath.indexOf('&*&')
-  if (firstIndex !== -1) {
-    newPath = newPath.substring(firstIndex + 3, path.length)
-  }
-  const index = newPath.lastIndexOf('.')
-  const fileType = newPath.substring(index, newPath.length)
-  const fileName = newPath.split('\\').pop() || tOriginal('MilkdownEditor.customFile.unknownName')
-  return { fileType, fileName }
-}
 interface CustomFileProps {
   type: UploadFileTypeProps
 }
