+++ app/renderer/src/main/src/components/pluginHasParamsDrawer/PluginHasParamsDrawer.tsx
@@ -13,7 +13,7 @@ import type { YakParamProps } from '@/pages/plugins/pluginsType'
 import { getJsonSchemaListResult } from '../JsonFormWrapper/JsonFormWrapper'
 import { failed } from '@/utils/notification'
 import { ExecuteEnterNodeByPluginParams } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/LocalPluginExecuteDetailHeard'
-import { ExtraParamsNodeByType } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/PluginExecuteExtraParams'
+import { ExtraParamsNodeByType } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/PluginExecuteExtraParamsNodes'
 import { useMemoizedFn } from 'ahooks'
 import type { YakitDrawerProps } from '../yakitUI/YakitDrawer/YakitDrawerType'
 import { useI18nNamespaces } from '@/i18n/useI18nNamespaces'
