+++ app/renderer/src/main/src/pages/ai-agent/historyChat/HistoryChat.tsx
@@ -16,7 +16,7 @@ import type { AIAgentTriggerEventInfo } from '../aiAgentType'
 import emiter from '@/utils/eventBus/eventBus'
 import { grpcDeleteAISession } from '../grpc'
 import type { AISession, DeleteAISessionRequest } from '../type/aiChat'
-import { SideSettingButton } from '../aiChatWelcome/AIChatWelcome'
+import { SideSettingButton } from '../aiChatWelcome/AIChatWelcomeSideSetting'
 import HistoryChatList, { DAY_MS, getChatTimestamp } from './HistoryChatList/HistoryChatList'
 import { useI18nNamespaces } from '@/i18n/useI18nNamespaces'
 import useSessionList from './HistoryChatList/hook/useSessionList'
