+++ app/renderer/src/main/src/pages/aiForge/forgeEditor/ForgeEditor.tsx
@@ -49,7 +49,7 @@ import type { GetAIToolListRequest } from '@/pages/ai-agent/type/aiTool'
 import { YakitRadioButtons } from '@/components/yakitUI/YakitRadioButtons/YakitRadioButtons'
 import { YakitEditor } from '@/components/yakitUI/YakitEditor/YakitEditor'
 import { ExecuteEnterNodeByPluginParams } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/LocalPluginExecuteDetailHeard'
-import { ExtraParamsNodeByType } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/PluginExecuteExtraParams'
+import { ExtraParamsNodeByType } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/PluginExecuteExtraParamsNodes'
 import type { YakParamProps } from '@/pages/plugins/pluginsType'
 import { getValueByType, onCodeToInfo, ParamsToGroupByGroupName } from '@/pages/plugins/editDetails/utils'
 import cloneDeep from 'lodash/cloneDeep'
