+++ app/renderer/src/main/src/pages/configManagement/HotPatchManagement.tsx
@@ -5,7 +5,7 @@ import { YakitInput } from '@/components/yakitUI/YakitInput/YakitInput'
 import { YakitEditor } from '@/components/yakitUI/YakitEditor/YakitEditor'
 import { YakitModal } from '@/components/yakitUI/YakitModal/YakitModal'
 import { yakitFailed, yakitNotify } from '@/utils/notification'
-import { type HotPatchTempItem, AddHotCodeTemplate, type HotCodeType } from '@/pages/fuzzer/HTTPFuzzerHotPatch'
+import { type HotPatchTempItem, AddHotCodeTemplate, type HotCodeType } from '@/pages/fuzzer/hotPatchShared'
 import { HotPatchDefaultContent, HotPatchTempDefault } from '@/defaultConstants/HTTPFuzzerPage'
 import { AnalyzeHotPatchTempDefault, MITMHotPatchTempDefault } from '@/defaultConstants/mitm'
 import { cloneDeep } from 'lodash'
