+++ app/renderer/src/main/src/pages/fuzzer/FuzzerSequence/FuzzerSequence.tsx
@@ -52,7 +52,8 @@ import { Divider, Result, Tooltip } from 'antd'
 import { YakitSelect } from '@/components/yakitUI/YakitSelect/YakitSelect'
 import { YakitPopover } from '@/components/yakitUI/YakitPopover/YakitPopover'
 import classNames from 'classnames'
-import { LabelNodeItem, MatcherAndExtractionDrawer } from '../MatcherAndExtractionCard/MatcherAndExtractionCard'
+import { LabelNodeItem } from '../MatcherAndExtractionCard/MatcherAndExtractionCardShared'
+import { MatcherAndExtractionDrawer } from '../MatcherAndExtractionCard/MatcherAndExtractionCard'
 import { YakitSwitch } from '@/components/yakitUI/YakitSwitch/YakitSwitch'
 import { YakitSpin } from '@/components/yakitUI/YakitSpin/YakitSpin'
 import { yakitFailed, yakitNotify } from '@/utils/notification'
@@ -125,7 +126,7 @@ import { useI18nNamespaces } from '@/i18n/useI18nNamespaces'
 import { AdvancedSetV2, ConcurrencyItem, initSetValue } from './FuzzerPageConcurrency'
 import { YakitDrawer } from '@/components/yakitUI/YakitDrawer/YakitDrawer'
 import { useProxy } from '@/hook/useProxy'
-import { HotCodeTemplate, type HotPatchTempItem } from '../HTTPFuzzerHotPatch'
+import { HotCodeTemplate, type HotPatchTempItem } from '../hotPatchShared'
 import { useGlobalHotPatchTag } from '@/store/globalHotPatch'
 
 const ResponseCard = React.lazy(() => import('./ResponseCard'))
