+++ app/renderer/src/main/src/pages/fuzzer/HttpQueryAdvancedConfig/FuzzerConfigPanels.tsx
@@ -11,7 +11,7 @@ import type {
   MatchingAndExtraction,
 } from '../MatcherAndExtractionCard/MatcherAndExtractionCardType'
 import { YakitRadioButtons } from '@/components/yakitUI/YakitRadioButtons/YakitRadioButtons'
-import { ExtractionResultsContent, onFilterEmptySubMatcher } from '../MatcherAndExtractionCard/MatcherAndExtractionCard'
+import { ExtractionResultsContent, onFilterEmptySubMatcher } from '../MatcherAndExtractionCard/MatcherAndExtractionCardShared'
 import { ExtractorsList, MatchersList } from './HttpQueryAdvancedConfig'
 import type { AdvancedConfigValueProps } from './HttpQueryAdvancedConfigType'
 import { StringToUint8Array } from '@/utils/str'
