+++ app/renderer/src/main/src/pages/fuzzer/MatcherAndExtractionCard/MatcherAndExtractionCard.tsx
@@ -73,6 +73,7 @@ import { useCampare } from '@/hook/useCompare/useCompare'
 import { type TFunction, useI18nNamespaces } from '@/i18n/useI18nNamespaces'
 import { Trans } from 'react-i18next'
 import i18n from '@/i18n/i18n'
+import { ExtractionResultsContent, LabelNodeItem, onFilterEmptySubMatcher } from './MatcherAndExtractionCardShared'
 import { YakitSegmented } from '@/components/yakitUI/YakitSegmented/YakitSegmented'
 import { SafeMarkdown } from '@/pages/assetViewer/reportRenders/markdownRender'
 
@@ -631,20 +632,6 @@ export const MatcherAndExtraction: React.FC<MatcherAndExtractionProps> = React.m
 
 const isMatcherEmpty = (matchersList) => {
   return matchersList?.filter((i) => !((i?.Group || []).map((i) => i.trim()).join('') === '')).length <= 0
-}
-export const onFilterEmptySubMatcher = (param: FilterEmptySubMatcherFunctionProps) => {
-  const { matchers, index, subIndex } = param
-  const matchersCopy = cloneDeep(matchers)
-  const newMatchers: HTTPResponseMatcher[] = []
-  matchersCopy.forEach((m, n) => {
-    if (n === index) {
-      m.SubMatchers = m.SubMatchers.filter((_, s) => s !== subIndex)
-    }
-    if (m.SubMatchers.length > 0) {
-      newMatchers.push(m)
-    }
-  })
-  return newMatchers
 }
 export const MatcherCollapse: React.FC<MatcherCollapseProps> = React.memo(
   forwardRef((props, ref) => {
@@ -1351,28 +1338,6 @@ export const ExtractorItem: React.FC<ExtractorItemProps> = React.memo((props) =>
   )
 })
 
-export const LabelNodeItem: React.FC<labelNodeItemProps> = React.memo((props) => {
-  const { column } = props
-  return (
-    <div
-      className={classNames(
-        styles['label-node'],
-        {
-          [styles['label-node-column']]: column,
-        },
-        props.className,
-      )}
-    >
-      <span
-        className={classNames(styles['label'], props.labelClassName)}
-        style={{ width: i18n.language.startsWith('zh') ? 104 : 130 }}
-      >
-        {props.label}
-      </span>
-      {props.children}
-    </div>
-  )
-})
 const colors: (t: TFunction) => { color: string; title: string }[] = (t) => {
   return [
     {
@@ -1471,21 +1436,6 @@ export const ColorSelect: React.FC<ColorSelectProps> = React.memo((props) => {
   )
 })
 
-export const ExtractionResultsContent: React.FC<ExtractionResultsContentProps> = React.memo((props) => {
-  const { list = [] } = props
-  return (
-    <div className={classNames(styles['extract-results'], 'yakit-descriptions')}>
-      <Descriptions bordered size="small" column={2}>
-        {list.map((item, index) => (
-          <Descriptions.Item label={<YakitCopyText showText={item.Key} />} key={`${item.Key}-${item.Value}`} span={2}>
-            {item.Value ? <YakitCopyText showText={item.Value} /> : ''}
-          </Descriptions.Item>
-        ))}
-      </Descriptions>
-    </div>
-  )
-})
-
 export const MatcherAndExtractionDrawer: React.FC<MatcherAndExtractionDrawerProps> = React.memo((props) => {
   const {
     pageType = 'webfuzzer',
@@ -1542,3 +1492,5 @@ export const MatcherAndExtractionDrawer: React.FC<MatcherAndExtractionDrawerProp
     </YakitDrawer>
   )
 })
+
+export { ExtractionResultsContent, LabelNodeItem, onFilterEmptySubMatcher } from './MatcherAndExtractionCardShared'
