+++ app/renderer/src/main/src/pages/fuzzer/components/HTTPFuzzerPageTable/HTTPFuzzerPageTable.tsx
@@ -29,7 +29,7 @@ import {
 import styles from './HTTPFuzzerPageTable.module.scss'
 import { HollowLightningBoltIcon } from '@/assets/newIcon'
 import { Alert, Divider, Tooltip } from 'antd'
-import { ExtractionResultsContent } from '../../MatcherAndExtractionCard/MatcherAndExtractionCard'
+import { ExtractionResultsContent } from '../../MatcherAndExtractionCard/MatcherAndExtractionCardShared'
 import { showYakitModal } from '@/components/yakitUI/YakitModal/YakitModalConfirm'
 import { YakitRadioButtons } from '@/components/yakitUI/YakitRadioButtons/YakitRadioButtons'
 import { YakitResizeBox } from '@/components/yakitUI/YakitResizeBox/YakitResizeBox'
@@ -39,7 +39,7 @@ import { openABSFileLocated, openExternalWebsite, openPacketNewWindow } from '@/
 import ReactResizeDetector from 'react-resize-detector'
 import { useCampare } from '@/hook/useCompare/useCompare'
 import { DefFuzzerTableMaxData } from '@/defaultConstants/HTTPFuzzerPage'
-import { CodingPopover } from '@/components/HTTPFlowDetail'
+import { CodingPopover } from '@/components/HTTPFlowDetailParts'
 import { OutlineSelectorIcon } from '@/assets/icon/outline'
 import { FuzzerRemoteGV } from '@/enums/fuzzer'
 import { isCellRedSingleColor } from '@/components/TableVirtualResize/utils'
