+++ app/renderer/src/main/src/pages/httpRequestBuilder/HTTPRequestBuilder.tsx
@@ -6,7 +6,7 @@ import { PlusIcon, TrashIcon } from '@/assets/newIcon'
 import { YakitButton } from '@/components/yakitUI/YakitButton/YakitButton'
 import styles from './HTTPRequestBuilder.module.scss'
 import type { HTTPRequestBuilderParams } from '@/models/HTTPRequestBuilder'
-import { SetVariableItem } from '../fuzzer/HttpQueryAdvancedConfig/HttpQueryAdvancedConfig'
+import { SetVariableItem } from '../fuzzer/HttpQueryAdvancedConfig/HttpQueryAdvancedConfigParts'
 import classNames from 'classnames'
 import { type PageNodeItemProps, usePageInfo } from '@/store/pageInfo'
 import { shallow } from 'zustand/shallow'
