+++ app/renderer/src/main/src/pages/pluginHub/pluginHubList/HubListOnline.tsx
@@ -48,10 +48,12 @@ import { SolidPluscircleIcon } from '@/assets/icon/solid'
 import emiter from '@/utils/eventBus/eventBus'
 import { YakitRoute } from '@/enums/yakitRoute'
 import {
+  YakitGetOnlinePlugin,
+} from '@/pages/mitm/MITMServerHijacking/MITMPluginOnline'
+import {
   PluginGroup,
   TagsAndGroupRender,
   type YakFilterRemoteObj,
-  YakitGetOnlinePlugin,
 } from '@/pages/mitm/MITMServerHijacking/MITMPluginLocalList'
 import { isCommunityEdition } from '@/utils/envfile'
 import { FilterPopoverBtn } from '@/pages/plugins/funcTemplate'
