+++ app/renderer/src/main/src/pages/pluginHub/pluginHubList/HubListOwn.tsx
@@ -43,7 +43,7 @@ import type { API } from '@/services/swagger/resposeType'
 import { SolidPluscircleIcon } from '@/assets/icon/solid'
 import emiter from '@/utils/eventBus/eventBus'
 import { YakitRoute } from '@/enums/yakitRoute'
-import { YakitGetOnlinePlugin } from '@/pages/mitm/MITMServerHijacking/MITMPluginLocalList'
+import { YakitGetOnlinePlugin } from '@/pages/mitm/MITMServerHijacking/MITMPluginOnline'
 import useGetSetState from '../hooks/useGetSetState'
 import { FilterPopoverBtn } from '@/pages/plugins/funcTemplate'
 import { Tooltip } from 'antd'
