+++ app/renderer/src/main/src/pages/yakRunnerAuditCode/AuditCode/AuditCode.tsx
@@ -27,7 +27,7 @@ import styles from './AuditCode.module.scss'
 import { genDefaultPagination, type QueryGeneralResponse, type YakScript } from '@/pages/invoker/schema'
 import { Divider, Form, Progress, Slider, Tooltip, Tree } from 'antd'
 import { YakitSpin } from '@/components/yakitUI/YakitSpin/YakitSpin'
-import { ExtraParamsNodeByType } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/PluginExecuteExtraParams'
+import { ExtraParamsNodeByType } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/PluginExecuteExtraParamsNodes'
 import { FormContentItemByType } from '@/pages/plugins/operator/localPluginExecuteDetailHeard/LocalPluginExecuteDetailHeard'
 import type { YakParamProps } from '@/pages/plugins/pluginsType'
 import { getValueByType, getYakExecutorParam, ParamsToGroupByGroupName } from '@/pages/plugins/editDetails/utils'
